licenta/requirements.txt
Django
django-crispy-forms
djangorestframework
pylint
pylint-django
autopep8
coverage
